test: extract logic in own package
